code/AnyhowCP-uai-2026-submission/pyproject.toml
[project]
name = "anyhowcp"
version = "0.2.0"
description = "Minimal research code for mask-valid conformal prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jax>=0.4",
    "numpy>=1.24",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.9",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
